refactored taskchatMessages
